requirements.txt
# Core dependencies
numpy>=1.20.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.0
networkx>=2.8.0
matplotlib>=3.5.0
nltk>=3.7
tqdm>=4.64.0
scipy>=1.7.0
concept-py>=0.9.0
python-dotenv>=1.0.0
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.5

# LLM Provider packages
openai>=1.0.0
google-generativeai>=0.4.0  # For Gemini models
anthropic>=0.8.0
requests>=2.0.0  # For Deepseek, Qwen, and general API calls
pydantic>=2.0.0  # For data validation
httpx>=0.25.0  # For async HTTP requests

# Async and Concurrency
anyio>=3.7.0
asyncio>=3.4.3

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-tornasync>=0.6.0
pytest-trio>=0.8.0
pytest-twisted>=1.14.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
# Core visualization dependencies
plotly>=5.0.0
networkx>=2.6.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Gradio for web interface
gradio>=4.0.0
streamlit
# Optional dependencies for advanced visualizations
# dash>=2.0.0  # For interactive dashboards
# jupyter-dash>=0.4.0  # For Jupyter notebook integration
# ipywidgets>=7.6.0  # For interactive widgets in Jupyter

